refactor(hickey): move tapBackdropAtSafePoint to support/drawer.ts
The helper's concept is "Corvu drawer backdrop tap" — shared across
mobile drawer step files. Homing it in mobile_drawer_steps.ts forced
mobile_code_browse_steps.ts to cross-import from a peer step file, an
authoring-order accident. Lift to support/drawer.ts so any step file
can import it without crossing peers.
